requirements.txt
xlrd==1.1.0
Django==2.0.6
django_filter==2.0.0
django_filters==0.2.1
python_magic==0.4.15
djangorestframework==3.9.0
django-oauth-toolkit==1.2.0
django-cors-headers==2.4.0    
django-cors-middleware==1.3.1
